backend/requirements.txt
openai
psycopg2-binary
fastapi
uvicorn
python-dotenv
sse-starlette
